requirements.txt
cryptography==43.0.3
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-graphql-jwt==0.4.0
django-jazzmin==3.0.1
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
pillow==11.0.0
psycopg2==2.9.10
python-decouple==3.8
django-cors-headers